stepper/pyproject.toml
[project]
name = "stepper"
version = "0.1.0"
description = "HackerFab photolithography stepper control"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.10",
    "pillow>=10.0",
    "pyserial>=3.5",
    "toml>=0.10",
    "ttkbootstrap>=1.10",
    "ultralytics>=8.3",
]

[project.optional-dependencies]
# Install with: uv sync --extra basler
basler = ["pypylon>=4.1"]

[tool.uv]
# This project is run as a script, not installed as a package.
package = false
